Use radio availability to sync card status instead of radio on
- In cases where RIL initializes more quickly than frameworks,
the card state changed notification is missed by frameworks.
To avoid missing card states, we register for radio on or
available. However, this is decided based on a property set
on encrypted devices.
- On undecrypted devices, we ALWAYS register for radio ON.
This breaks the user case where if we boot up in airplane
mode, the card state is never queried and the device
thinks there is no SIM.
- Make UiccController sync the card state on radio available
in all cases

Change-Id: I369e8faabde3baa6f834a511812d24eb0b9de340
CRs-Fixed: 881465