SubscriptionInfoUpdater: improved sim swap
We now keep track of the last number of inserted SIMs updated for. So
when the number of SIMs has changed we do the following:

 1 sim: Clear stale defaults
        Disable SMS prompt, SMS app will default to current sim
        Disable data if the SIM left in the tray is not the one chosen

 m sim: Enable SMS prompt if we don't have a valid default (aka always
        ask)

Furthermore, when we detect the device is an MSIM device and there is
only one active sub, return the active sub to the UI layers (so SIM
selection screen displays proper, grayed out, current settings).

Ticket: CYNGNOS-2185, CYNGNOS-2570

Change-Id: Ib1b4f4d87640e4b20722295304c58dff7dbeecbb
Signed-off-by: Roman Birg <roman@cyngn.com>
Signed-off-by: bgos10 <bishalgoswami@gmail.com>