Fix incorrect call radio type on UI during silent redial
When handover takes place, the ImsPhoneConnection objects
attached to the handover call are not cleaned up until the call
ends on the CS pipe after handover. As a result, the stale object
keeps updating UI on the IMS registration state. This causes the
call UI to show WiFi call when the call is on CS. The fix here is
to clean up the ImsPhoneConnection objects as soon as they are
mapped to CS connections upon handover.

Change-Id: I3d6f0a78a625394eeb4c5f24f9ab4c8d63783d10
CRs-Fixed: 999393